migrate project to npm
- aligned project with `actions/typescript-action` config
- bumped dependencies
