+++ .eslintignore
@@ -1,4 +1,6 @@
 .git
 node_modules
 dist
+lib
 coverage
+jest.config.js
