--- .eslintrc.js
@@ -1,29 +1,0 @@
-module.exports = {
-  parser: '@typescript-eslint/parser',
-  parserOptions: {
-    ecmaVersion: 9,
-    sourceType: 'module',
-    project: ['./tsconfig.json', './tsconfig.eslint.json'],
-  },
-
-  extends: ['airbnb-typescript/base', 'plugin:prettier/recommended', 'plugin:jest/recommended'],
-  rules: {
-    'import/named': 'off',
-    'import/prefer-default-export': 'off',
-    'no-console': 'off',
-    'no-restricted-globals': 'off',
-    'react/jsx-curly-newline': 'off',
-    'react/jsx-one-expression-per-line': 'off',
-    'react/jsx-props-no-spreading': 'off',
-    'react/jsx-wrap-multilines': 'off',
-    '@typescript-eslint/camelcase': 'off',
-    '@typescript-eslint/explicit-function-return-type': 'off',
-    '@typescript-eslint/indent': 'off',
-    '@typescript-eslint/quotes': 'off',
-  },
-  env: {
-    node: true,
-    es6: true,
-    'jest/globals': true,
-  },
-};
