--- .github/workflows/ci.yml
@@ -1,116 +1,0 @@
-name: CI
-
-on:
-  push:
-  pull_request:
-    types:
-      - opened
-      - synchronize
-
-jobs:
-
-  lint:
-    name: Lint
-    runs-on: ubuntu-latest
-    if: github.event_name == 'push'
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@master
-
-      - name: Install dependencies
-        run: yarn install --frozen-lockfile
-
-      - name: Lint
-        run: yarn lint
-
-  test:
-    name: Test
-    runs-on: ubuntu-latest
-    if: github.event_name == 'push'
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@master
-
-      - name: Install dependencies
-        run: yarn install --frozen-lockfile
-
-      - name: Test
-        run: yarn test
-        env:
-          CI: true
-
-  testAction:
-    name: Test action
-    runs-on: ubuntu-latest
-    if: github.event_name == 'push'
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@master
-
-      - name: Test action
-        uses: ./
-        with:
-          github-token: ${{ secrets.GITHUB_TOKEN }}
-          netlify-auth-token: ${{ secrets.NETLIFY_AUTH_TOKEN }}
-          netlify-site-id: ${{ secrets.NETLIFY_SITE_ID }}
-          build-dir: './example'
-          comment-on-pull-request: true
-          comment-on-commit: true
-          dry-run: true
-
-  deployDraft:
-    name: Deploy draft to Netlify
-    runs-on: ubuntu-latest
-    if: github.event_name == 'pull_request' && github.ref != 'refs/heads/master'
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v1
-
-      - name: Install 
-        run: yarn install --frozen-lockfile
-
-      - name: Build action
-        run: yarn build
-
-      - name: Generate simple HTML document
-        run: |
-          mkdir -p example
-          echo -e "<pre>$(date -u)\n$GITHUB_SHA\n$GITHUB_REF</pre>" > example/index.html
-
-      - name: Test action
-        uses: ./
-        with:
-          github-token: ${{ secrets.GITHUB_TOKEN }}
-          netlify-auth-token: ${{ secrets.NETLIFY_AUTH_TOKEN }}
-          netlify-site-id: ${{ secrets.NETLIFY_SITE_ID }}
-          build-dir: './example'
-          draft: true
-          comment-on-pull-request: true
-
-  deployProduction:
-    name: Deploy production to Netlify
-    runs-on: ubuntu-latest
-    if: github.event_name == 'push' && github.ref == 'refs/heads/master'
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v1
-
-      - name: Install 
-        run: yarn install --frozen-lockfile
-
-      - name: Build action
-        run: yarn build
-
-      - name: Generate simple HTML document
-        run: |
-          mkdir -p example
-          echo -e "<pre>$(date -u)\n$GITHUB_SHA\n$GITHUB_REF</pre>" > example/index.html
-
-      - name: Test action
-        uses: ./
-        with:
-          github-token: ${{ secrets.GITHUB_TOKEN }}
-          netlify-auth-token: ${{ secrets.NETLIFY_AUTH_TOKEN }}
-          netlify-site-id: ${{ secrets.NETLIFY_SITE_ID }}
-          build-dir: './example'
-          comment-on-commit: true
