+++ .prettierignore
@@ -1,4 +1,5 @@
 .git
 node_modules
 dist
+lib
 coverage
