+++ jest.config.js
@@ -2,7 +2,8 @@ module.exports = {
   clearMocks: true,
   moduleFileExtensions: ['js', 'ts'],
   testEnvironment: 'node',
-  testMatch: ['**/*.spec.ts'],
+  testMatch: ['**/*.test.ts'],
+  testRunner: 'jest-circus/runner',
   transform: {
     '^.+\\.ts$': 'ts-jest',
   },
