+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@south-paw/action-netlify-deploy",
-  "version": "1.0.4",
-  "description": "Deploys a build to Netlify and can comment on the commit or PR with the URL",
+  "version": "1.1.0",
+  "description": "Netlify deployments via GitHub actions",
   "keywords": [],
   "homepage": "https://github.com/South-Paw/action-netlify-deploy",
   "bugs": "https://github.com/South-Paw/action-netlify-deploy/issues",
@@ -11,45 +11,41 @@
     "email": "hello@southpaw.co.nz",
     "url": "http://southpaw.co.nz/"
   },
-  "files": [
-    "dist/"
-  ],
-  "main": "dist/index.js",
+  "main": "lib/main.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/South-Paw/action-netlify-deploy.git"
   },
   "scripts": {
-    "build": "ncc build src/index.ts && rimraf dist/typescript",
-    "clean": "rimraf dist",
-    "lint": "yarn lint:type-check && yarn lint:eslint",
+    "clean": "rimraf dist lib",
+    "build:tsc": "tsc",
+    "build:ncc": "ncc build",
+    "build": "npm run clean && npm run build:tsc && npm run build:ncc",
+    "lint": "npm run lint:type-check && npm run lint:eslint",
     "lint:eslint": "eslint \"src/**/*.{js,ts}\"",
     "lint:type-check": "tsc --noEmit",
     "test": "jest",
     "test:coverage": "jest --coverage"
   },
   "dependencies": {
-    "@actions/core": "^1.2.4",
-    "@actions/github": "^2.2.0",
-    "netlify": "^4.1.7"
+    "@actions/core": "^1.2.6",
+    "@actions/github": "^4.0.0",
+    "netlify": "^6.1.9"
   },
   "devDependencies": {
-    "@types/jest": "^25.2.3",
-    "@types/node": "^14.0.5",
-    "@typescript-eslint/eslint-plugin": "^3.0.1",
-    "@typescript-eslint/parser": "^3.0.1",
-    "@zeit/ncc": "^0.22.3",
-    "eslint": "^7.1.0",
-    "eslint-config-airbnb-typescript": "^7.2.1",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-import": "^2.20.2",
-    "eslint-plugin-jest": "^23.13.2",
-    "eslint-plugin-prettier": "^3.1.3",
-    "jest": "^26.0.1",
-    "prettier": "^2.0.5",
+    "@types/jest": "^26.0.20",
+    "@types/node": "^14.14.31",
+    "@typescript-eslint/parser": "^4.15.1",
+    "@vercel/ncc": "^0.25.1",
+    "eslint": "^7.20.0",
+    "eslint-plugin-github": "^4.1.1",
+    "eslint-plugin-jest": "^24.1.5",
+    "jest": "^26.6.3",
+    "jest-circus": "^26.6.3",
+    "prettier": "2.2.1",
     "rimraf": "^3.0.2",
-    "ts-jest": "^26.0.0",
-    "typescript": "^3.9.3"
+    "ts-jest": "^26.5.1",
+    "typescript": "^4.1.5"
   },
   "private": true
 }
